requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boolean.py==5.0
Brotli==1.1.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cssselect2==0.8.0
curl_cffi==0.12.0
cycler==0.12.1
cyclonedx-python-lib==9.1.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.19.1
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.0
fqdn==1.5.1
frozendict==2.4.6
funcparserlib==1.0.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
holidays==0.78
httpcore==1.0.9
httpx==0.28.1
hy==1.1.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lark==1.2.2
license-expression==30.4.4
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.3
msgpack==1.1.1
multitasking==0.0.12
narwhals==2.1.0
nest-asyncio==1.6.0
numpy==2.3.2
ordered-set==4.1.0
overrides==7.7.0
packageurl-python==0.17.5
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
peewee==3.18.2
pillow==11.3.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
processor==0.10.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-serializable==2.1.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.3
pyphen==0.17.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==13.9.4
rpds-py==0.26.0
scipy==1.16.1
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
Twiggy==0.4.7
twiggy-goodies==0.12.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uri-template==1.3.0
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
zopfli==0.2.3.post1
tensorflow
opencv-python-headless